+++ blockstream/index.js
@@ -1,3 +1,4 @@
+const getBlockstreamTx = require('./get_blockstream_tx');
 const getBlockstreamVout = require('./get_blockstream_vout');
 
-module.exports = {getBlockstreamVout};
+module.exports = {getBlockstreamTx, getBlockstreamVout};